with pic 16F887
The pic 16F887 microprocessor is programmed. MQ4 methane gas
sensor, potentiometer is used because I could not find the library of the weight sensor in the proteus. received from sensors.
After the data is processed in the pic card, the outputs obtained are displayed to the users with the LCD.
transferred.